# swagger-output.json
{
    "openapi":  "3.0.4",
    "info":  {
                 "title":  "AI-PMS.API",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/AISuggestion":  {
                                            "get":  {
                                                        "tags":  [
                                                                     "AISuggestion"
                                                                 ],
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "OK"
                                                                                  }
                                                                      }
                                                    },
                                            "post":  {
                                                         "tags":  [
                                                                      "AISuggestion"
                                                                  ],
                                                         "requestBody":  {
                                                                             "content":  {
                                                                                             "application/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/AISuggestionDto"
                                                                                                                                 }
                                                                                                                  },
                                                                                             "text/json":  {
                                                                                                               "schema":  {
                                                                                                                              "$ref":  "#/components/schemas/AISuggestionDto"
                                                                                                                          }
                                                                                                           },
                                                                                             "application/*+json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/AISuggestionDto"
                                                                                                                                   }
                                                                                                                    }
                                                                                         }
                                                                         },
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK"
                                                                                   }
                                                                       }
                                                     }
                                        },
                  "/api/AISuggestion/{id}":  {
                                                 "get":  {
                                                             "tags":  [
                                                                          "AISuggestion"
                                                                      ],
                                                             "parameters":  [
                                                                                {
                                                                                    "name":  "id",
                                                                                    "in":  "path",
                                                                                    "required":  true,
                                                                                    "schema":  {
                                                                                                   "type":  "string",
                                                                                                   "format":  "uuid"
                                                                                               }
                                                                                }
                                                                            ],
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK"
                                                                                       }
                                                                           }
                                                         }
                                             },
                  "/api/AISuggestion/{id}/read":  {
                                                      "put":  {
                                                                  "tags":  [
                                                                               "AISuggestion"
                                                                           ],
                                                                  "parameters":  [
                                                                                     {
                                                                                         "name":  "id",
                                                                                         "in":  "path",
                                                                                         "required":  true,
                                                                                         "schema":  {
                                                                                                        "type":  "string",
                                                                                                        "format":  "uuid"
                                                                                                    }
                                                                                     }
                                                                                 ],
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  "OK"
                                                                                            }
                                                                                }
                                                              }
                                                  },
...
                  "/api/BottleneckDetection":  {
                                                   "post":  {
                                                                "tags":  [
                                                                             "BottleneckDetection"
                                                                         ],
                                                                "requestBody":  {
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/BottleneckRequest"
                                                                                                                                        }
                                                                                                                         },
                                                                                                    "text/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/BottleneckRequest"
                                                                                                                                 }
                                                                                                                  },
                                                                                                    "application/*+json":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/BottleneckRequest"
                                                                                                                                          }
                                                                                                                           }
                                                                                                }
                                                                                },
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "OK",
                                                                                              "content":  {
                                                                                                              "text/plain":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/BottleneckResponse"
                                                                                                                                            }
                                                                                                                             },
                                                                                                              "application/json":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/BottleneckResponse"
                                                                                                                                                  }
                                                                                                                                   },
                                                                                                              "text/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/BottleneckResponse"
                                                                                                                                           }
                                                                                                                            }
                                                                                                          }
                                                                                          }
                                                                              }
                                                            }
                                               },
                  "/api/BottleneckDetection/project/{projectId}":  {
                                                                       "post":  {
                                                                                    "tags":  [
                                                                                                 "BottleneckDetection"
                                                                                             ],
                                                                                    "parameters":  [
                                                                                                       {
                                                                                                           "name":  "projectId",
                                                                                                           "in":  "path",
                                                                                                           "required":  true,
                                                                                                           "schema":  {
                                                                                                                          "type":  "string",
                                                                                                                          "format":  "uuid"
                                                                                                                      }
                                                                                                       }
                                                                                                   ],
                                                                                    "responses":  {
                                                                                                      "200":  {
                                                                                                                  "description":  "OK",
                                                                                                                  "content":  {
                                                                                                                                  "text/plain":  {
                                                                                                                                                     "schema":  {
                                                                                                                                                                    "$ref":  "#/components/schemas/BottleneckResponse"
                                                                                                                                                                }
                                                                                                                                                 },
                                                                                                                                  "application/json":  {
                                                                                                                                                           "schema":  {
                                                                                                                                                                          "$ref":  "#/components/schemas/BottleneckResponse"
                                                                                                                                                                      }
                                                                                                                                                       },
                                                                                                                                  "text/json":  {
                                                                                                                                                    "schema":  {
                                                                                                                                                                   "$ref":  "#/components/schemas/BottleneckResponse"
                                                                                                                                                               }
                                                                                                                                                }
                                                                                                                              }
                                                                                                              }
                                                                                                  }
                                                                                }
                                                                   },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "AISuggestionDto":  {
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "id":  {
                                                                                             "type":  "string",
                                                                                             "format":  "uuid"
                                                                                         },
                                                                                  "projectId":  {
                                                                                                    "type":  "string",
                                                                                                    "format":  "uuid",
                                                                                                    "nullable":  true
                                                                                                },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "nullable":  true
                                                                                           },
                                                                                  "title":  {
                                                                                                "type":  "string",
                                                                                                "nullable":  true
                                                                                            },
                                                                                  "description":  {
                                                                                                      "type":  "string",
                                                                                                      "nullable":  true
                                                                                                  },
                                                                                  "priority":  {
                                                                                                   "type":  "string",
                                                                                                   "nullable":  true
                                                                                               },
                                                                                  "isRead":  {
                                                                                                 "type":  "boolean"
                                                                                             },
                                                                                  "createdAt":  {
                                                                                                    "type":  "string",
                                                                                                    "format":  "date-time"
                                                                                                }
                                                                              },
                                                               "additionalProperties":  false
                                                           },
...
                                       "BottleneckItem":  {
                                                              "type":  "object",
                                                              "properties":  {
                                                                                 "taskTitle":  {
                                                                                                   "type":  "string",
                                                                                                   "nullable":  true
                                                                                               },
                                                                                 "reason":  {
                                                                                                "type":  "string",
                                                                                                "nullable":  true
                                                                                            },
                                                                                 "severity":  {
                                                                                                  "type":  "string",
                                                                                                  "nullable":  true
                                                                                              },
                                                                                 "suggestedActions":  {
                                                                                                          "type":  "array",
                                                                                                          "items":  {
                                                                                                                        "type":  "string"
                                                                                                                    },
                                                                                                          "nullable":  true
                                                                                                      }
                                                                             },
                                                              "additionalProperties":  false
                                                          },
                                       "BottleneckRequest":  {
                                                                 "type":  "object",
                                                                 "properties":  {
                                                                                    "projectName":  {
                                                                                                        "type":  "string",
                                                                                                        "nullable":  true
                                                                                                    },
                                                                                    "tasks":  {
                                                                                                  "type":  "array",
                                                                                                  "items":  {
                                                                                                                "$ref":  "#/components/schemas/BottleneckTaskDto"
                                                                                                            },
                                                                                                  "nullable":  true
                                                                                              }
                                                                                },
                                                                 "additionalProperties":  false
                                                             },
                                       "BottleneckResponse":  {
                                                                  "type":  "object",
                                                                  "properties":  {
                                                                                     "bottlenecks":  {
                                                                                                         "type":  "array",
                                                                                                         "items":  {
                                                                                                                       "$ref":  "#/components/schemas/BottleneckItem"
                                                                                                                   },
                                                                                                         "nullable":  true
                                                                                                     },
                                                                                     "recommendations":  {
                                                                                                             "type":  "array",
                                                                                                             "items":  {
                                                                                                                           "type":  "string"
                                                                                                                       },
                                                                                                             "nullable":  true
                                                                                                         }
                                                                                 },
                                                                  "additionalProperties":  false
                                                              },
                                       "BottleneckTaskDto":  {
                                                                 "type":  "object",
                                                                 "properties":  {
                                                                                    "title":  {
                                                                                                  "type":  "string",
                                                                                                  "nullable":  true
                                                                                              },
                                                                                    "status":  {
                                                                                                   "type":  "string",
                                                                                                   "nullable":  true
                                                                                               },
                                                                                    "complexity":  {
                                                                                                       "type":  "string",
                                                                                                       "nullable":  true
                                                                                                   },
                                                                                    "estimatedHours":  {
                                                                                                           "type":  "integer",
                                                                                                           "format":  "int32"
                                                                                                       },
                                                                                    "assignedTo":  {
                                                                                                       "type":  "string",
                                                                                                       "nullable":  true
                                                                                                   }
                                                                                },
                                                                 "additionalProperties":  false
                                                             },
...
                                   }
                   },
...
}